LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/start-edge-in-block-layout-direction-expected.txt



PASS Horizontal LTR.
PASS Horizontal RTL.
PASS Vertical-LR LTR.
PASS Vertical-LR RTL.
FAIL Vertical-RL LTR. assert_equals: expected -120 but got -200
FAIL Vertical-RL RTL. assert_equals: expected -120 but got -150
